added new branch, divided tests into user and book
